Extract separate files for clases
