+++ cg_lims/EPPs/qc/sequencing_quality_control.py
@@ -3,6 +3,7 @@ import sys
 import click
 
 from cg_lims.EPPs.qc.sequencing_quality_checker import SequencingQualityChecker
+from cg_lims.EPPs.qc.sequencing_artifact_manager import SequencingArtifactManager
 
 
 @click.command()
@@ -11,9 +12,13 @@ def sequencing_quality_control(ctx):
     process = ctx.obj["process"]
     status_db_api = ctx.obj["status_db_api"]
 
-    quality_checker: SequencingQualityChecker = SequencingQualityChecker(
-        process=process, status_db_api=status_db_api
+    artifact_manager = SequencingArtifactManager(process)
+
+    quality_checker = SequencingQualityChecker(
+        artifact_manager=artifact_manager,
+        status_db_api=status_db_api,
     )
+
     quality_summary: str = quality_checker.validate_sequencing_quality()
 
     if quality_checker.samples_failed_quality_control():
